Refactor: 사용자 회원가입 테스트 클래스 이름 변경
